fix(agents-api): Fix blob-store interceptor implementation
